End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { MessageParams } from '../_models/message-params';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,10 @@ export class MessagesComponent implements OnInit {
   messageParams: MessageParams;
   loading: boolean = false;
 
-  constructor(private messageService: MessageService) {
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService
+  ) {
     this.messageParams = this.messageService.getMessageParams();
   }
 
@@ -37,12 +41,22 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: () => {
-        const messageIndex = this.messages.findIndex((m) => m.id === id);
-        this.messages.splice(messageIndex, 1);
-      },
-    });
+    this.confirmService
+      .confirm('Confirm delete message', 'This cannot be undone')
+      .subscribe({
+        next: (result) => {
+          if (result) {
+            this.messageService.deleteMessage(id).subscribe({
+              next: () => {
+                const messageIndex = this.messages.findIndex(
+                  (m) => m.id === id
+                );
+                this.messages.splice(messageIndex, 1);
+              },
+            });
+          }
+        },
+      });
   }
 
   pageChanged(event: any) {
